requirements.txt
django 
djangorestframework 
psycopg2-binary 
django-environ 
drf-spectacular
djangorestframework-simplejwt 
django-cors-headers
Pillow